// frontend/src/api/generated/models/stravaArchiveImportStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type StravaArchiveImportStatus = typeof StravaArchiveImportStatus[keyof typeof StravaArchiveImportStatus];


export const StravaArchiveImportStatus = {
  Pending: 'Pending',
  Downloading: 'Downloading',
  Analyzing: 'Analyzing',
  PreviewReady: 'PreviewReady',
  Importing: 'Importing',
  Succeeded: 'Succeeded',
  Failed: 'Failed',
  Cancelled: 'Cancelled',
} as const;